refactor: remove un-necessary event sender
