+++ pubspec.yaml
@@ -13,7 +13,7 @@ dependencies:
   ffi: ^2.2.0
   hooks: ^1.0.0
   native_toolchain_rust: ^1.0.3
-  audiopc_interface: ^0.0.1+2
+  audiopc_interface: ^0.0.1+3
 
 dev_dependencies:
   flutter_lints: ^6.0.0
